Make better use of the cache

+++ build_install_chapel.sh
@@ -1,13 +1,15 @@
 #!/bin/bash
-cd 
-wget https://github.com/chapel-lang/chapel/releases/download/1.16.0/chapel-1.16.0-1.tar.gz
-tar xzf chapel-1.16.0-1.tar.gz
-rm chapel-1.16.0-1.tar.gz
-cd chapel-1.16.0
-source util/setchplenv.bash
-make
-export CHPL_AUX_FILESYS=curl
-patch -p0 <<EOF
+cd || exit
+if [[ ! -d ~/chapel-1.16.0 ]]
+then
+    wget https://github.com/chapel-lang/chapel/releases/download/1.16.0/chapel-1.16.0-1.tar.gz
+    tar xzf chapel-1.16.0-1.tar.gz
+    rm chapel-1.16.0-1.tar.gz
+    cd chapel-1.16.0 || exit
+    source util/setchplenv.bash
+    make
+    export CHPL_AUX_FILESYS=curl
+    patch -p0 <<EOF
 --- modules/packages/Curl.chpl  2017-10-10 17:45:03.000000000 -0400
 +++ modules/packages/Curl-new.chpl  2017-10-13 13:03:02.000000000 -0400
 @@ -299,18 +299,18 @@
@@ -32,4 +34,5 @@ patch -p0 <<EOF
    return true;
  }
 EOF
-make
+    make
+fi
